Rename column image and done (Customer and Empoyee) login function, Display personal informaition on customer and Employees.

+++ src/main/java/com/gentlemonster/GentleMonsterBE/Services/User/UserService.java
@@ -7,22 +7,32 @@ import com.gentlemonster.GentleMonsterBE.DTO.Requests.User.UserRequest;
 import com.gentlemonster.GentleMonsterBE.DTO.Responses.APIResponse;
 import com.gentlemonster.GentleMonsterBE.DTO.Responses.PagingResponse;
 import com.gentlemonster.GentleMonsterBE.DTO.Responses.User.BaseUserResponse;
+import com.gentlemonster.GentleMonsterBE.DTO.Responses.User.UserInforResponse;
 import com.gentlemonster.GentleMonsterBE.DTO.Responses.User.UserResponse;
 import com.gentlemonster.GentleMonsterBE.Entities.Address;
+import com.gentlemonster.GentleMonsterBE.Entities.AuthToken;
 import com.gentlemonster.GentleMonsterBE.Entities.Media;
 import com.gentlemonster.GentleMonsterBE.Entities.Role;
 import com.gentlemonster.GentleMonsterBE.Entities.Store;
 import com.gentlemonster.GentleMonsterBE.Entities.User;
+import com.gentlemonster.GentleMonsterBE.Exception.NotFoundException;
 import com.gentlemonster.GentleMonsterBE.Repositories.IAddressRepository;
 import com.gentlemonster.GentleMonsterBE.Repositories.IRoleRepository;
 import com.gentlemonster.GentleMonsterBE.Repositories.IStoreRepository;
+import com.gentlemonster.GentleMonsterBE.Repositories.ITokenRepository;
 import com.gentlemonster.GentleMonsterBE.Repositories.IUserRepository;
 import com.gentlemonster.GentleMonsterBE.Repositories.Specification.UserSpecification;
 import com.gentlemonster.GentleMonsterBE.Services.Cloudinary.CloudinaryService;
-import com.gentlemonster.GentleMonsterBE.Utils.LocalizationUtil;
-import com.gentlemonster.GentleMonsterBE.Utils.VietnameseStringUtils;
+import com.gentlemonster.GentleMonsterBE.Utils.JwtTokenUtils;
+import com.gentlemonster.GentleMonsterBE.Utils.LocalizationUtils;
+import com.gentlemonster.GentleMonsterBE.Utils.ValidationUtils;
+
+import io.jsonwebtoken.JwtException;
 import jakarta.transaction.Transactional;
 import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -40,8 +50,10 @@ import java.text.SimpleDateFormat;
 import java.time.LocalDateTime;
 import java.util.*;
 
+
 @Service
-@NoArgsConstructor
+@RequiredArgsConstructor
+@Slf4j
 public class UserService implements IUserService {
     @Autowired
     private IUserRepository iUserRepository;
@@ -52,15 +64,19 @@ public class UserService implements IUserService {
     @Autowired
     private ModelMapper modelMapper;
     @Autowired
-    private LocalizationUtil localizationUtil;
+    private LocalizationUtils localizationUtil;
     @Autowired
     private PasswordEncoder passwordEncode;
     @Autowired
     private IAddressRepository iAddressRepository;
     @Autowired
-    private VietnameseStringUtils vietnameseStringUtils;
+    private ValidationUtils vietnameseStringUtils;
     @Autowired
     private CloudinaryService cloudinaryService;
+    @Autowired
+    private ITokenRepository iTokenRepository;
+    @Autowired
+    private JwtTokenUtils jwtTokenUtils;
 
     @Override
     public PagingResponse<List<BaseUserResponse>> getAllUser(@ModelAttribute UserRequest userRequest) {
@@ -94,33 +110,6 @@ public class UserService implements IUserService {
         return new PagingResponse<>(userResponseList,messages,userPage.getTotalPages(),userPage.getTotalElements());
     }
 
-//    @Override
-//    public PagingResponse<List<BaseUserResponse>> getAllUser(@ModelAttribute UserRequest userRequest) {
-//        List<BaseUserResponse> userResponseList;
-//        List<User> userList;
-//        Pageable pageable;
-//
-//        int page = Math.max(userRequest.getPage() - 1, 0); // Page index should start from 0
-//        int size = userRequest.getLimit() > 0 ? userRequest.getLimit() : 10; // Default size is 10 if limit is not provided
-//        pageable = PageRequest.of(page, size, Sort.by("fullName").descending());
-//
-//        Page<User> userPage = iUserRepository.findAll(pageable);
-//        userList = userPage.getContent();
-//
-//        // Map the User entities to UserResponse DTOs
-//        userResponseList = userList.stream()
-//                .map(user -> modelMapper.map(user, BaseUserResponse.class))
-//                .toList();
-//
-//        if(userResponseList.isEmpty()){
-//            return new PagingResponse<>(null, List.of(localizationUtil.getLocalizedMessage(MessageKey.USER_NOT_EXIST)), 1, 0L);
-//        }
-//        // Prepare the response
-//        List<String> messages = new ArrayList<>();
-//        messages.add(localizationUtil.getLocalizedMessage(MessageKey.USER_GET_SUCCESS));
-//        return new PagingResponse<>(userResponseList, messages, userPage.getTotalPages(), userPage.getTotalElements());
-//    }
-
     @Override
     public PagingResponse<List<UserResponse>> getAllUserByRole(UserRequest userRequest) {
         List<UserResponse> userResponseList;
@@ -131,7 +120,7 @@ public class UserService implements IUserService {
         int size = userRequest.getLimit() > 0 ? userRequest.getLimit() : 10; // Default size is 10 if limit is not provided
         pageable = PageRequest.of(page, size, Sort.by("fullName").descending());
 
-        Specification<User> specification = UserSpecification.filterUsers(userRequest.getRole(), userRequest.getName(), userRequest.getEmail());
+        // Specification<User> specification = UserSpecification.filterUsers(userRequest.getRole(), userRequest.getName(), userRequest.getEmail());
         Page<User> userPage = iUserRepository.findAll(pageable);
         userList = userPage.getContent();
 
@@ -151,7 +140,7 @@ public class UserService implements IUserService {
 
     @Override
     @Transactional
-    public APIResponse<Boolean> addUser(AddUserResquest addUserRequest) {
+    public APIResponse<Boolean> addUser(AddUserResquest addUserRequest) throws NotFoundException {
         if(iUserRepository.existsByEmail(addUserRequest.getEmail())){
             return new APIResponse<>(null,List.of(localizationUtil.getLocalizedMessage(MessageKey.USER_ALREADY_EXIST)));
         }
@@ -209,19 +198,20 @@ public class UserService implements IUserService {
         Role role = iRoleRepository.findByName(addUserRequest.getRole()).orElse(null);
         user.setRole(role);
         if (role == null) {
-            return new APIResponse<>(null, List.of(localizationUtil.getLocalizedMessage("Role not found")));
+            // return new APIResponse<>(null, List.of(localizationUtil.getLocalizedMessage("Role not found")));
+            throw new NotFoundException(localizationUtil.getLocalizedMessage(MessageKey.ROLE_NOT_FOUND));
         }
         // Tạo mã code dựa trên role
         user.setCode(generateCode(role));
-        if(addUserRequest.getRole().equals("STORE_MANAGER") || addUserRequest.getRole().equals("EMPLOYEE")){
-        Store userStore = iStoreRepository.findByStoreName(addUserRequest.getStore()).orElse(null);
-        if (userStore == null) {
-            return new APIResponse<>(null, List.of(localizationUtil.getLocalizedMessage(MessageKey.STORE_WRONG_ROLE)));
-        }
-        user.setStore(userStore);
-        }else {
-            iUserRepository.save(user);
-        }
+        // if(addUserRequest.getRole().equals("STORE_MANAGER") || addUserRequest.getRole().equals("EMPLOYEE")){
+        // Store userStore = iStoreRepository.findByStoreName(addUserRequest.getStore()).orElse(null);
+        // if (userStore == null) {
+        //     return new APIResponse<>(null, List.of(localizationUtil.getLocalizedMessage(MessageKey.STORE_WRONG_ROLE)));
+        // }
+        // user.setStore(userStore);
+        // }else {
+        //     iUserRepository.save(user);
+        // }
         iUserRepository.save(user);
         List<String> messages = new ArrayList<>();
         messages.add(localizationUtil.getLocalizedMessage(MessageKey.USER_CREATE_SUCCESS));
@@ -229,10 +219,11 @@ public class UserService implements IUserService {
     }
 
     @Override
-    public APIResponse<UserResponse> getOneUser(String userID) {
+    public APIResponse<UserResponse> getOneUser(String userID) throws NotFoundException {
         User user = iUserRepository.findById(UUID.fromString(userID)).orElse(null);
         if(user == null){
-            return new APIResponse<>(null,List.of(localizationUtil.getLocalizedMessage(MessageKey.USER_NOT_EXIST)));
+            // return new APIResponse<>(null,List.of(localizationUtil.getLocalizedMessage(MessageKey.USER_NOT_FOUND)));
+            throw new NotFoundException(localizationUtil.getLocalizedMessage(MessageKey.USER_NOT_FOUND));
         }
         UserResponse userResponse = modelMapper.map(user,UserResponse.class);
         List<String> messages = new ArrayList<>();
@@ -241,14 +232,16 @@ public class UserService implements IUserService {
     }
 
     @Override
-    public APIResponse<Boolean> updateUser(String userID, EditUserRequest editUserRequest) {
+    public APIResponse<Boolean> updateUser(String userID, EditUserRequest editUserRequest) throws NotFoundException {
         User user = iUserRepository.findById(UUID.fromString(userID)).orElse(null);
         Role role = iRoleRepository.findByName(editUserRequest.getRole()).orElse(null);
         if (user == null){
-            return new APIResponse<>(null,List.of(localizationUtil.getLocalizedMessage(MessageKey.USER_NOT_EXIST)));
+            // return new APIResponse<>(null,List.of(localizationUtil.getLocalizedMessage(MessageKey.USER_NOT_FOUND)));
+            throw new NotFoundException(localizationUtil.getLocalizedMessage(MessageKey.USER_NOT_FOUND));
         }
         if (role == null){
-            return new APIResponse<>(null,List.of(localizationUtil.getLocalizedMessage(MessageKey.ROLE_NOT_EXIST)));
+            // return new APIResponse<>(null,List.of(localizationUtil.getLocalizedMessage(MessageKey.ROLE_NOT_)));
+            throw new NotFoundException(localizationUtil.getLocalizedMessage(MessageKey.ROLE_NOT_FOUND));
         }
 
         modelMapper.map(editUserRequest,user);
@@ -283,28 +276,40 @@ public class UserService implements IUserService {
                 .build();
         user.getAddresses().add(address);
         user.setModifiedDate(LocalDateTime.now());
-        if(editUserRequest.getRole().equals("STORE_MANAGER") || editUserRequest.getRole().equals("EMPLOYEE")){
-            Store userStore = iStoreRepository.findByStoreName(editUserRequest.getStore()).orElse(null);
-            if (userStore == null) {
-                return new APIResponse<>(null, List.of(localizationUtil.getLocalizedMessage(MessageKey.STORE_WRONG_ROLE)));
-            }
-            user.setStore(userStore);
-        }else {
-            iUserRepository.save(user);
-        }
+        // if(editUserRequest.getRole().equals("STORE_MANAGER") || editUserRequest.getRole().equals("EMPLOYEE")){
+        //     Store userStore = iStoreRepository.findByStoreName(editUserRequest.getStore()).orElse(null);
+        //     if (userStore == null) {
+        //         return new APIResponse<>(null, List.of(localizationUtil.getLocalizedMessage(MessageKey.STORE_WRONG_ROLE)));
+        //     }
+        //     user.setStore(userStore);
+        // }else {
+        //     iUserRepository.save(user);
+        // }
         iUserRepository.save(user);
-//        UserResponse userResponse = modelMapper.map(user,UserResponse.class);
+        // UserResponse userResponse = modelMapper.map(user,UserResponse.class);
         List<String> messages = new ArrayList<>();
         messages.add(localizationUtil.getLocalizedMessage(MessageKey.USER_UPDATE_SUCCESS));
         return new APIResponse<>(true,messages);
     }
 
     @Override
-    public APIResponse<Boolean> deleteUser(String userID) {
-        User user = iUserRepository.findById(UUID.fromString(userID)).orElseThrow(() -> new RuntimeException("User not found"));
-//        if(user.getRole().getName().equals("ADMIN")){
-//            return new APIResponse<>(false,List.of(localizationUtil.getLocalizedMessage(MessageKey.NOT_DELETE_ADMIN_USER)));
-//        }
+    public APIResponse<Boolean> deleteUser(String userID) throws NotFoundException {
+        User user = iUserRepository.findById(UUID.fromString(userID)).orElse(null);
+        if (user == null) {
+            // return new APIResponse<>(null, List.of(localizationUtil.getLocalizedMessage(MessageKey.USER_NOT_FOUND)));
+            throw new NotFoundException(localizationUtil.getLocalizedMessage(MessageKey.USER_NOT_FOUND));
+        }
+        Media avatar = user.getImage();
+        if (avatar != null && avatar.getPublicId() != null) {
+            // Xóa media thumbnail nếu có
+            cloudinaryService.deleteMedia(avatar.getPublicId());
+        }
+        AuthToken authToken = iTokenRepository.findByUserId(user.getId()).orElse(null);
+        if (authToken != null) {
+            // Xóa token liên quan đến người dùng
+            iTokenRepository.delete(authToken);
+        }
+
         iUserRepository.delete(user);
         List<String> messages = new ArrayList<>();
         messages.add(localizationUtil.getLocalizedMessage(MessageKey.USER_DELETE_SUCCESS));
@@ -397,10 +402,10 @@ public class UserService implements IUserService {
     }
 
     @Override
-    public APIResponse<Boolean> uploadAvatarEmployee(String userID, MultipartFile image) {
+    public APIResponse<Boolean> uploadAvatarEmployee(String userID, MultipartFile image) throws NotFoundException {
         User user = iUserRepository.findById(UUID.fromString(userID)).orElse(null);
         if (user == null) {
-            return new APIResponse<>(null, List.of(localizationUtil.getLocalizedMessage(MessageKey.USER_NOT_EXIST)));
+            throw new NotFoundException(localizationUtil.getLocalizedMessage(MessageKey.USER_NOT_FOUND));
         }
         try {
             Map uploadResult = cloudinaryService.uploadMedia(image, "users");
@@ -413,7 +418,7 @@ public class UserService implements IUserService {
                     .altText("Avatar: " + user.getFullName())
                     .type("IMAGE")
                     .build();
-            user.setAvatar(media);
+            user.setImage(media);
             iUserRepository.save(user);
             List<String> messages = new ArrayList<>();
             messages.add(localizationUtil.getLocalizedMessage(MessageKey.USER_UPLOAD_AVATAR_SUCCESS));
@@ -423,5 +428,41 @@ public class UserService implements IUserService {
             return new APIResponse<>(null, List.of(localizationUtil.getLocalizedMessage(MessageKey.USER_UPLOAD_AVATAR_FAILED)));
         }
 
+    }
+
+    @Override
+    public UserInforResponse getUserLoginResponse(String login) throws NotFoundException {
+        if (login == null || !login.startsWith("Bearer ")) {
+            log.error("Invalid or missing Authorization header");
+            throw new IllegalArgumentException(localizationUtil.getLocalizedMessage(MessageKey.TOKEN_INVALID));
+        }
+        // Lấy token từ chuỗi login
+        String token = login.substring(7).trim();
+        if (token == null || token.isEmpty()) {
+            log.error("Token is null or empty");
+            throw new IllegalArgumentException(localizationUtil.getLocalizedMessage(MessageKey.TOKEN_INVALID));
+        }
+        if (jwtTokenUtils.isTokenExpired(token)) {
+            log.error("Token has expired");
+            throw new IllegalArgumentException(localizationUtil.getLocalizedMessage(MessageKey.TOKEN_EXPIRED));
+        }  
+        String subject = jwtTokenUtils.extractSubject(token);
+        log.info("Extracted subject: {}, role: {}", subject, jwtTokenUtils.extractClaim(token, claims -> claims.get("role", String.class)));
+        if (token == null || token.isEmpty()) {
+            log.error("Invalid or missing token");
+            throw new IllegalArgumentException(localizationUtil.getLocalizedMessage(MessageKey.TOKEN_INVALID));
+        }
+        Optional<User> userOptional;
+        if (jwtTokenUtils.extractClaim(token, claims -> claims.get("role", String.class)).equals("CUSTOMER")) {
+            userOptional = iUserRepository.findByEmail(subject);
+        } else {
+            userOptional = iUserRepository.findByUsername(subject);
+        }
+        UserInforResponse userResponse = modelMapper.map(userOptional, UserInforResponse.class);
+        if (userResponse == null) {
+            log.error("User not found for token: {}", token);
+            throw new NotFoundException(localizationUtil.getLocalizedMessage(MessageKey.USER_NOT_FOUND));
+        }
+        return userResponse;
     }
 }
